try forcing bfloat16

+++ trl/trainer/rloo_trainer.py
@@ -81,7 +81,7 @@ class RLOOTrainer(PolicyTrainerBase):
 
         context_length = queries.shape[1]
 
-        with torch.cuda.amp.autocast():
+        with torch.cuda.amp.autocast(dtype=torch.bfloat16):
             with torch.no_grad():
                 query_responses, logits = self.generate(
                     model,
